GH-2149 all SPARQL query tests now use parametrized setup
