+++ compliance/sparql/src/test/java/org/eclipse/rdf4j/query/parser/sparql/manifest/W3CApprovedSPARQL11QueryTest.java
@@ -19,6 +19,7 @@ import org.junit.Ignore;
 import junit.framework.Test;
 
 @Ignore("replaced by org.eclipse.rdf4j.sail.memory.MemorySPARQL11QueryComplianceTest")
+@Deprecated
 public class W3CApprovedSPARQL11QueryTest extends SPARQLQueryTest {
 
 	public static Test suite() throws Exception {
@@ -47,7 +48,7 @@ public class W3CApprovedSPARQL11QueryTest extends SPARQLQueryTest {
 						"sq03 - Subquery within graph pattern, graph variable is not bound",
 						// test case is incorrect wrt SPARQL 1.1 spec, see https://github.com/eclipse/rdf4j/issues/1978
 //						"agg empty group",
-//						"Aggregate over empty group resulting in a row with unbound variables" 
+//						"Aggregate over empty group resulting in a row with unbound variables"
 				};
 
 				return new W3CApprovedSPARQL11QueryTest(testURI, name, queryFileURL, resultFileURL, dataSet,
