+++ compliance/sparql/src/test/java/org/eclipse/rdf4j/sail/federation/FederationSPARQL10QueryTest.java
@@ -8,57 +8,17 @@
 package org.eclipse.rdf4j.sail.federation;
 
 import org.eclipse.rdf4j.query.Dataset;
-import org.eclipse.rdf4j.query.parser.sparql.manifest.SPARQL10ManifestTest;
-import org.eclipse.rdf4j.query.parser.sparql.manifest.SPARQLQueryTest;
+import org.eclipse.rdf4j.query.parser.sparql.manifest.SPARQL10QueryComplianceTest;
 import org.eclipse.rdf4j.repository.Repository;
 import org.eclipse.rdf4j.repository.dataset.DatasetRepository;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.eclipse.rdf4j.sail.memory.MemoryStore;
 
-import junit.framework.Test;
+public class FederationSPARQL10QueryTest extends SPARQL10QueryComplianceTest {
 
-public class FederationSPARQL10QueryTest extends SPARQLQueryTest {
-
-	public static Test suite() throws Exception {
-		return SPARQL10ManifestTest.suite(new Factory() {
-
-			@Override
-			public FederationSPARQL10QueryTest createSPARQLQueryTest(String testURI, String name, String queryFileURL,
-					String resultFileURL, Dataset dataSet, boolean laxCardinality) {
-				return createSPARQLQueryTest(testURI, name, queryFileURL, resultFileURL, dataSet, laxCardinality,
-						false);
-			}
-
-			@Override
-			public FederationSPARQL10QueryTest createSPARQLQueryTest(String testURI, String name, String queryFileURL,
-					String resultFileURL, Dataset dataSet, boolean laxCardinality, boolean checkOrder) {
-				String[] ignoredTests = {
-						// incompatible with SPARQL 1.1 - syntax for decimals was modified
-						"Basic - Term 6",
-						// incompatible with SPARQL 1.1 - syntax for decimals was modified
-						"Basic - Term 7",
-						// Test is incorrect: assumes timezoned date is comparable with non-timezoned
-						"date-2",
-						// fails on federated sail for unknown reason - ignoring for now as non-reproducible in separate
-						// setup. See GH-2140
-						"Join scope - 1"
-				};
-
-				return new FederationSPARQL10QueryTest(testURI, name, queryFileURL, resultFileURL, dataSet,
-						laxCardinality, checkOrder, ignoredTests);
-			}
-		});
-
-	}
-
-	protected FederationSPARQL10QueryTest(String testURI, String name, String queryFileURL, String resultFileURL,
-			Dataset dataSet, boolean laxCardinality, String... ignoredTests) {
-		this(testURI, name, queryFileURL, resultFileURL, dataSet, laxCardinality, false, ignoredTests);
-	}
-
-	protected FederationSPARQL10QueryTest(String testURI, String name, String queryFileURL, String resultFileURL,
-			Dataset dataSet, boolean laxCardinality, boolean checkOrder, String... ignoredTests) {
-		super(testURI, name, queryFileURL, resultFileURL, dataSet, laxCardinality, checkOrder, ignoredTests);
+	public FederationSPARQL10QueryTest(String displayName, String testURI, String name, String queryFileURL,
+			String resultFileURL, Dataset dataset, boolean ordered) {
+		super(displayName, testURI, name, queryFileURL, resultFileURL, dataset, ordered);
 	}
 
 	@Override
