+++ compliance/sparql/src/test/java/org/eclipse/rdf4j/sail/nativerdf/NativeSPARQL11QueryComplianceTest.java
@@ -5,29 +5,33 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/org/documents/edl-v10.php.
  *******************************************************************************/
-package org.eclipse.rdf4j.query.parser.sparql.manifest;
+package org.eclipse.rdf4j.sail.nativerdf;
 
 import org.eclipse.rdf4j.query.Dataset;
+import org.eclipse.rdf4j.query.parser.sparql.manifest.SPARQL11QueryComplianceTest;
 import org.eclipse.rdf4j.repository.Repository;
 import org.eclipse.rdf4j.repository.dataset.DatasetRepository;
 import org.eclipse.rdf4j.repository.sail.SailRepository;
-import org.eclipse.rdf4j.sail.memory.MemoryStore;
+import org.junit.Rule;
+import org.junit.rules.TemporaryFolder;
 
 /**
- * SPARQL Query tests using testcases from the SPARQL 1.2 Working Group
- *
- * @author Jeen Broekstra
+ * @author jeen
  *
  */
-public class SPARQL12QueryTest extends SPARQL12QueryComplianceTest {
+public class NativeSPARQL11QueryComplianceTest extends SPARQL11QueryComplianceTest {
 
-	public SPARQL12QueryTest(String displayName, String testURI, String name, String queryFileURL, String resultFileURL,
-			Dataset dataset, boolean ordered) {
+	public NativeSPARQL11QueryComplianceTest(String displayName, String testURI, String name, String queryFileURL,
+			String resultFileURL, Dataset dataset, boolean ordered) {
 		super(displayName, testURI, name, queryFileURL, resultFileURL, dataset, ordered);
 	}
 
+	@Rule
+	public TemporaryFolder folder = new TemporaryFolder();
+
 	@Override
-	protected Repository newRepository() {
-		return new DatasetRepository(new SailRepository(new MemoryStore()));
+	protected Repository newRepository() throws Exception {
+		return new DatasetRepository(new SailRepository(new NativeStore(folder.newFolder(), "spoc")));
 	}
+
 }
